Enum dprint_swc_ecma_ast_view::Expr

pub enum Expr<'a> {

    This(&'a ThisExpr<'a>),
    Array(&'a ArrayLit<'a>),
    Object(&'a ObjectLit<'a>),
    Fn(&'a FnExpr<'a>),
    Unary(&'a UnaryExpr<'a>),
    Update(&'a UpdateExpr<'a>),
    Bin(&'a BinExpr<'a>),
    Assign(&'a AssignExpr<'a>),
    Member(&'a MemberExpr<'a>),
    Cond(&'a CondExpr<'a>),
    Call(&'a CallExpr<'a>),
    New(&'a NewExpr<'a>),
    Seq(&'a SeqExpr<'a>),
    Ident(&'a Ident<'a>),
    Lit(Lit<'a>),
    Tpl(&'a Tpl<'a>),
    TaggedTpl(&'a TaggedTpl<'a>),
    Arrow(&'a ArrowExpr<'a>),
    Class(&'a ClassExpr<'a>),
    Yield(&'a YieldExpr<'a>),
    MetaProp(&'a MetaPropExpr<'a>),
    Await(&'a AwaitExpr<'a>),
    Paren(&'a ParenExpr<'a>),
    JSXMember(&'a JSXMemberExpr<'a>),
    JSXNamespacedName(&'a JSXNamespacedName<'a>),
    JSXEmpty(&'a JSXEmptyExpr<'a>),
    JSXElement(&'a JSXElement<'a>),
    JSXFragment(&'a JSXFragment<'a>),
    TsTypeAssertion(&'a TsTypeAssertion<'a>),
    TsConstAssertion(&'a TsConstAssertion<'a>),
    TsNonNull(&'a TsNonNullExpr<'a>),
    TsAs(&'a TsAsExpr<'a>),
    PrivateName(&'a PrivateName<'a>),
    OptChain(&'a OptChainExpr<'a>),
    Invalid(&'a Invalid<'a>),

}

Variants

This(&'a ThisExpr<'a>)

Tuple Fields

0: &'a ThisExpr<'a>

Array(&'a ArrayLit<'a>)

Tuple Fields

0: &'a ArrayLit<'a>

Object(&'a ObjectLit<'a>)

Tuple Fields

0: &'a ObjectLit<'a>

Fn(&'a FnExpr<'a>)

Tuple Fields

0: &'a FnExpr<'a>

Unary(&'a UnaryExpr<'a>)

Tuple Fields

0: &'a UnaryExpr<'a>

Update(&'a UpdateExpr<'a>)

Tuple Fields

0: &'a UpdateExpr<'a>

++v, --v, v++, v--

Bin(&'a BinExpr<'a>)

Tuple Fields

0: &'a BinExpr<'a>

Assign(&'a AssignExpr<'a>)

Tuple Fields

0: &'a AssignExpr<'a>

Member(&'a MemberExpr<'a>)

Tuple Fields

0: &'a MemberExpr<'a>

A member expression. If computed is true, the node corresponds to a computed (a[b]) member expression and property is an Expression. If computed is false, the node corresponds to a static (a.b) member expression and property is an Identifier.

Cond(&'a CondExpr<'a>)

Tuple Fields

0: &'a CondExpr<'a>

true ? ‘a’ : ‘b’

Call(&'a CallExpr<'a>)

Tuple Fields

0: &'a CallExpr<'a>

New(&'a NewExpr<'a>)

Tuple Fields

0: &'a NewExpr<'a>

new Cat()

Seq(&'a SeqExpr<'a>)

Tuple Fields

0: &'a SeqExpr<'a>

Ident(&'a Ident<'a>)

Tuple Fields

0: &'a Ident<'a>

Lit(Lit<'a>)

Tuple Fields

0: Lit<'a>

Tpl(&'a Tpl<'a>)

Tuple Fields

0: &'a Tpl<'a>

TaggedTpl(&'a TaggedTpl<'a>)

Tuple Fields

0: &'a TaggedTpl<'a>

Arrow(&'a ArrowExpr<'a>)

Tuple Fields

0: &'a ArrowExpr<'a>

Class(&'a ClassExpr<'a>)

Tuple Fields

0: &'a ClassExpr<'a>

Yield(&'a YieldExpr<'a>)

Tuple Fields

0: &'a YieldExpr<'a>

MetaProp(&'a MetaPropExpr<'a>)

Tuple Fields

0: &'a MetaPropExpr<'a>

Await(&'a AwaitExpr<'a>)

Tuple Fields

0: &'a AwaitExpr<'a>

Paren(&'a ParenExpr<'a>)

Tuple Fields

0: &'a ParenExpr<'a>

JSXMember(&'a JSXMemberExpr<'a>)

Tuple Fields

0: &'a JSXMemberExpr<'a>

JSXNamespacedName(&'a JSXNamespacedName<'a>)

Tuple Fields

0: &'a JSXNamespacedName<'a>

JSXEmpty(&'a JSXEmptyExpr<'a>)

Tuple Fields

0: &'a JSXEmptyExpr<'a>

JSXElement(&'a JSXElement<'a>)

Tuple Fields

0: &'a JSXElement<'a>

JSXFragment(&'a JSXFragment<'a>)

Tuple Fields

0: &'a JSXFragment<'a>

TsTypeAssertion(&'a TsTypeAssertion<'a>)

Tuple Fields

0: &'a TsTypeAssertion<'a>

TsConstAssertion(&'a TsConstAssertion<'a>)

Tuple Fields

0: &'a TsConstAssertion<'a>

TsNonNull(&'a TsNonNullExpr<'a>)

Tuple Fields

0: &'a TsNonNullExpr<'a>

TsAs(&'a TsAsExpr<'a>)

Tuple Fields

0: &'a TsAsExpr<'a>

PrivateName(&'a PrivateName<'a>)

Tuple Fields

0: &'a PrivateName<'a>

OptChain(&'a OptChainExpr<'a>)

Tuple Fields

0: &'a OptChainExpr<'a>

Invalid(&'a Invalid<'a>)

Tuple Fields

0: &'a Invalid<'a>

Implementations

impl<'a> Expr<'a>

pub fn to<T: CastableNode<'a>>(&self) -> Option<&'a T>

pub fn expect<T: CastableNode<'a>>(&self) -> &'a T

pub fn is<T: CastableNode<'a>>(&self) -> bool