Trait dprint_swc_ecma_ast_view::NodeTrait

pub trait NodeTrait<'a>: SpannedExt {

    fn parent(&self) -> Option<Node<'a>>;

    fn children(&self) -> Vec<Node<'a>>;

    fn into_node(&self) -> Node<'a>;

    fn kind(&self) -> NodeKind;

    fn ancestors(&self) -> AncestorIterator<'a> { ... }

    fn start_line(&self) -> usize { ... }

    fn end_line(&self) -> usize { ... }

    fn start_column(&self) -> usize { ... }

    fn end_column(&self) -> usize { ... }

    fn width(&self) -> usize { ... }

    fn child_index(&self) -> usize { ... }

    fn previous_sibling(&self) -> Option<Node<'a>> { ... }

    fn previous_siblings(&self) -> Vec<Node<'a>> { ... }

    fn next_sibling(&self) -> Option<Node<'a>> { ... }

    fn next_siblings(&self) -> Vec<Node<'a>> { ... }

    fn tokens(&self) -> &'a [TokenAndSpan] { ... }

    fn children_with_tokens(&self) -> Vec<NodeOrToken<'a>> { ... }

    fn children_with_tokens_fast(
        &self, 
        program: &dyn RootNode<'a>
    ) -> Vec<NodeOrToken<'a>> { ... }

    fn leading_comments(&self) -> CommentsIterator<'a> { ... }

    fn trailing_comments(&self) -> CommentsIterator<'a> { ... }

    fn program(&self) -> Program<'a> { ... }

    fn module(&self) -> &Module<'a> { ... }

    fn script(&self) -> &Script<'a> { ... }

    fn text(&self) -> &'a str { ... }

    fn previous_token(&self) -> Option<&'a TokenAndSpan> { ... }

    fn next_token(&self) -> Option<&'a TokenAndSpan> { ... }

    fn previous_tokens(&self) -> &'a [TokenAndSpan] { ... }

    fn next_tokens(&self) -> &'a [TokenAndSpan] { ... }

}

Required methods

fn parent(&self) -> Option<Node<'a>>

fn children(&self) -> Vec<Node<'a>>

fn into_node(&self) -> Node<'a>

fn kind(&self) -> NodeKind

Loading content...
Provided methods

fn ancestors(&self) -> AncestorIterator<'a>

fn start_line(&self) -> usize

fn end_line(&self) -> usize

fn start_column(&self) -> usize

fn end_column(&self) -> usize

fn width(&self) -> usize

fn child_index(&self) -> usize

fn previous_sibling(&self) -> Option<Node<'a>>

fn previous_siblings(&self) -> Vec<Node<'a>>

Gets the previous siblings in the order they appear in the file.

fn next_sibling(&self) -> Option<Node<'a>>

Gets the next siblings in the order they appear in the file.

fn next_siblings(&self) -> Vec<Node<'a>>

fn tokens(&self) -> &'a [TokenAndSpan]

fn children_with_tokens(&self) -> Vec<NodeOrToken<'a>>

fn children_with_tokens_fast(
&self,
program: &dyn RootNode<'a>
) -> Vec<NodeOrToken<'a>>

fn leading_comments(&self) -> CommentsIterator<'a>

fn trailing_comments(&self) -> CommentsIterator<'a>

fn program(&self) -> Program<'a>

Gets the root node.

fn module(&self) -> &Module<'a>

Gets the root node if the view was created from a Module; otherwise panics.

fn script(&self) -> &Script<'a>

Gets the root node if the view was created from a Script; otherwise panics.

fn text(&self) -> &'a str

fn previous_token(&self) -> Option<&'a TokenAndSpan>

fn next_token(&self) -> Option<&'a TokenAndSpan>

fn previous_tokens(&self) -> &'a [TokenAndSpan]

Gets the previous tokens in the order they appear in the file.

fn next_tokens(&self) -> &'a [TokenAndSpan]

Gets the next tokens in the order they appear in the file.

Loading content...